+++ lib/cart_stepper.dart
@@ -1,3 +1,4 @@
 library cart_stepper;
 
 export 'src/cart_stepper.dart';
+export 'src/stepper_style.dart';